[#98] style: move user projection directory
